update: show menu entry or commands and man pages better:
On installing a package, show its menu entry (if any), with
nice spacing, in bold text.

If no menu entry, list the commands available and any man pages.

Includes a fix for finding the file containing the paths to the
.desktop files (which contain the menu entries)

Also:

If updating system cache takes more than 2 seconds, pop up an
extra message, reading "Updating cache..", so user knows what
is going on. (install audacity-data to see it in action.)