fix: add pkgbinarymangler to blacklisted packages:
In fossapup64, "pkgbinarymangler_144" replaces dpkg-deb with
a shell script.. This breaks Pkg, which uses dpkg-deb during
installation of packages.

So blacklist pkgbinarymangler, so it never gets installed.

Note, this package is of no real interest in Puppy Linux
anyway - it is used to re-doing maintainer info during
building of Debian/Ubuntu.